Add UX response for checkbox success
Also do a tiny bit of refactoring and cleanup
